Added helpMessage. Added jacked up. Improved saved game
Jacked up means that if the king is on the jack level, and if the opposing team uses a jack to rub the bottom of the deck then the king team goes back down to 2.

Improved save game to include and constantly be saving the score, meaning which team is playing which team, who is the current king and what number they are playing on.